+++ packages/valory/services/contribution/service.yaml
@@ -8,10 +8,10 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeiaeysyipptsieavfpyfkrfl3n37klu3ux7o2rtks2w4k2nnr2hshu
 fingerprint_ignore_patterns: []
-agent: valory/contribution:0.1.0:bafybeido7dgarvv6mwt3u4v26ynbgtvk5lfiqm5gjjtgj74izrrp453xpi
+agent: valory/contribution:0.1.0:bafybeif67ugh66olrqicugjcsgkhljz56f6m5ee33keqdxwyb2dphzkbx4
 number_of_agents: 4
 ---
-public_id: valory/contribution_skill_abci:0.1.0
+public_id: valory/contribution_abci:0.1.0
 type: skill
 models:
   params:
