+++ packages/valory/skills/contribution_abci/fsm_specification.yaml
