+++ packages/valory/skills/contribution_abci/skill.yaml
@@ -1,4 +1,4 @@
-name: contribution_skill_abci
+name: contribution_abci
 author: valory
 version: 0.1.0
 type: skill
@@ -6,18 +6,18 @@ description: The chained abci for the contribution skill.
 license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
-  __init__.py: bafybeihrdde2hozs5v5nwwwf3vvinneyxtdyni4bbqlxp7r222sqajlaai
-  behaviours.py: bafybeicezipgftkdnvvsccc7pdlwdbhxoddz4ojphaqtu52jfhsvmjreoq
+  __init__.py: bafybeig6mpu2fw7r5ymwc4whcqdl3tygcpykz3jm7d5m4pc5l562mjcigi
+  behaviours.py: bafybeiddpsijxf2qa5wvzofbeq3uxuj6c2bm7dz2bfp557mbcybortyjpa
   composition.py: bafybeiakkvfphtv2kp5qh4tqhfebzeabsyw7ffc7dooucvmtxhb44vdqym
   dialogues.py: bafybeiflpquascygrfxcgb4ufiigqykmvbxlgbrxxa6cxmiy6f4w3aknmq
   fsm_specification.yaml: bafybeiepsgr4oekk52gpkw6g46cqzs7bklr45glj7h2xgas4tki765m6h4
   handlers.py: bafybeidaowt474fmhewix74bmb4ewp5xksxzugntrwnhpmtzjokeysr3tm
-  models.py: bafybeig33yrjdds7r2ek7qzddaqxnfn4o42lspbl275bihf6ec2k22y5au
+  models.py: bafybeidzplk4nxxjiwqfihmlfe2gqwbrkd2ft3qk2hvbwuwb4qpm6mswku
   tests/__init__.py: bafybeicukkp7ajzvqtmpedzwo23azuorvy5qtftadsxbwmv5zhk2tktjc4
-  tests/test_behaviours.py: bafybeigvr3eugdknpa7id4nem4ln73vhm4jgf2soayc2qsmt66fs24ctzm
-  tests/test_dialogues.py: bafybeigbzt4ran7uhltd7bkcjcyjzjj4l2dm4w2bw3bbvqx573bv4g5jzu
-  tests/test_handlers.py: bafybeiaomqulgu76t37hqdb5cjqbl6gqj7epekadbxjmoiszgcuzbeujsi
-  tests/test_models.py: bafybeiaxapvzgj44jnd6vautsex47w5npsoznadvvwb4isph55qgbz7rua
+  tests/test_behaviours.py: bafybeiali637pe2hkgmb4js2xf3qootkebmo73o6yfcynydmursvngtoze
+  tests/test_dialogues.py: bafybeic7hvuqahi4n3otym4ffe5d3pchv2jvlfeqff4xhsnpo376c5b3t4
+  tests/test_handlers.py: bafybeidejb25zv26pi5efdiis3utxk3riligtbej6eidkekwx67waiyqgy
+  tests/test_models.py: bafybeihm4cut6rvxpojdjb2ff4ez2vomqyer72ottvpvgkvahyrmmu673m
 fingerprint_ignore_patterns: []
 connections: []
 contracts: []
