+++ packages/valory/skills/contribution_abci/tests/test_dialogues.py
@@ -17,17 +17,10 @@
 #
 # ------------------------------------------------------------------------------
 
-"""Test the models.py module of the contribution skill."""
+"""Test the dialogues.py module of the contribution skill."""
 
-from packages.valory.skills.contribution_skill_abci.models import SharedState
-from packages.valory.skills.abstract_round_abci.test_tools.base import DummyContext
+import packages.valory.skills.contribution_abci.dialogues  # pylint: disable=unused-import # noqa: F401
 
 
-class TestSharedState:  # pylint: disable=too-few-public-methods
-    """Test SharedState of the contribution skill."""
-
-    def test_initialization(  # pylint: disable=no-self-use
-        self,
-    ) -> None:
-        """Test initialization."""
-        SharedState(name="", skill_context=DummyContext())
+def test_import() -> None:
+    """Test that the 'dialogues.py' of the contribution skill can be imported."""
